poboljsane perfomanse ako nema avatar

+++ src/backend/Controllers/PostController.cs
@@ -511,48 +511,51 @@ namespace backend.Controllers
                     message = "Error"
                 });
             var user = await _context.Users.FirstOrDefaultAsync(u => u.Username == User.Identity.Name);
-            Byte[] b = System.IO.File.ReadAllBytes(user.Avatar);
-            var streamAvatar = new MemoryStream(b);
-            using var _x = picture.OpenReadStream();
-
-            var fileStreamContent = new StreamContent(_x);
-            fileStreamContent.Headers.ContentType = new MediaTypeHeaderValue("image/*");
-
-            var multipartFormContent = new MultipartFormDataContent();
-            multipartFormContent.Add(fileStreamContent, name: "pictures", fileName: picture.FileName);
-            fileStreamContent = new StreamContent(streamAvatar);
-            multipartFormContent.Add(fileStreamContent, name: "pictures", fileName: "slika.jpeg");
-
-            var url = _configuration.GetSection("Microservice").Value + "/compare";
-
-            string responseString = null;
-            try 
+            if (user.Avatar.Contains("default.png") == false)
             {
-                var response = await _client.PostAsync(url, multipartFormContent);
-                response.EnsureSuccessStatusCode();
-                responseString = await response.Content.ReadAsStringAsync();
-            }
-            catch(HttpRequestException e)
-            {
-                Console.WriteLine("\nException Caught!");	
-                Console.WriteLine("Message :{0} ", e.Message);
+                Byte[] b = System.IO.File.ReadAllBytes(user.Avatar);
+                var streamAvatar = new MemoryStream(b);
+                using var _x = picture.OpenReadStream();
 
-                return BadRequest(new
+                var fileStreamContent = new StreamContent(_x);
+                fileStreamContent.Headers.ContentType = new MediaTypeHeaderValue("image/*");
+
+                var multipartFormContent = new MultipartFormDataContent();
+                multipartFormContent.Add(fileStreamContent, name: "pictures", fileName: picture.FileName);
+                fileStreamContent = new StreamContent(streamAvatar);
+                multipartFormContent.Add(fileStreamContent, name: "pictures", fileName: "slika.jpeg");
+
+                var url = _configuration.GetSection("Microservice").Value + "/compare";
+
+                string responseString = null;
+                try
                 {
-                    error=true,
-                    message="Error with microservice"
-                });
-            }
-
-            if (responseString.Contains("true"))
-            {
-                _context.Posts.Remove(post);
-                await _context.SaveChangesAsync();
-                return Ok(new
+                    var response = await _client.PostAsync(url, multipartFormContent);
+                    response.EnsureSuccessStatusCode();
+                    responseString = await response.Content.ReadAsStringAsync();
+                }
+                catch (HttpRequestException e)
                 {
-                    error = true,
-                    message = "Don't post a picture of yourself"
-                });
+                    Console.WriteLine("\nException Caught!");
+                    Console.WriteLine("Message :{0} ", e.Message);
+
+                    return BadRequest(new
+                    {
+                        error = true,
+                        message = "Error with microservice"
+                    });
+                }
+
+                if (responseString.Contains("true"))
+                {
+                    _context.Posts.Remove(post);
+                    await _context.SaveChangesAsync();
+                    return Ok(new
+                    {
+                        error = true,
+                        message = "Don't post a picture of yourself"
+                    });
+                }
             }
 
             string path = CreatePathToDataRoot(post.Id, picture.FileName);
